requirements.txt
streamlit
gtts
playsound==1.2.2
pyttsx3
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community 
sentence-transformers
